Refactor: join display modules in package
* Simplify support, when amount of code will raise
* Do not deprecate:
  `display_common`, `curses_display` and `raw_display` imports
* `web` display: extract javascript and css into separate files:
  simplify maintainability
