+++ urwid/display/_web.css
@@ -1,0 +1,12 @@
+body {    margin: 8px 8px 8px 8px; border: 0;
+    color: black; background-color: silver;
+    font-family: fixed; overflow: hidden; }
+
+form { margin: 0 0 8px 0; }
+
+#text { position: relative;
+    background-color: silver;
+    width: 100%; height: 100%;
+    margin: 3px 0 0 0; border: 1px solid #999; }
+
+#page { position: relative;  width: 100%;height: 100%;}
