+++ urwid/font.py
@@ -25,7 +25,7 @@ import warnings
 from pprint import pformat
 
 from urwid.canvas import CanvasError, TextCanvas
-from urwid.escape import SAFE_ASCII_DEC_SPECIAL_RE
+from urwid.display.escape import SAFE_ASCII_DEC_SPECIAL_RE
 from urwid.util import apply_target_encoding, str_util
 
 if typing.TYPE_CHECKING:
